+++ src/MusicLibraryApi.Client/ReleaseNotes.md
@@ -1,2 +1,5 @@
-﻿**1.0.0** / 2019.12.16**
+﻿**2.0.0** / 2019.12.16**
+* Removed playbacks fields from input song contract.
+
+**1.0.0** / 2019.12.16**
 * Initial version.